+++ src/main/java/uk/ac/ebi/eva/contigalias/controller/contigalias/ContigAliasController.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.ac.ebi.eva.contigalias.controller;
+package uk.ac.ebi.eva.contigalias.controller.contigalias;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
